requirements.txt
python-dotenv
sqlalchemy
pyodbc
pandas
fastapi
uvicorn